+++ cpm/__init__.py
@@ -1,5 +1,6 @@
 __author__ = "Lenard Dome"
 __email__ = "lenarddome@gmail.com"
+__license__ = "AGPL-3.0"
 
 """
 This is the main module of the package.
@@ -18,7 +19,7 @@ from . import models
 from . import applications
 from . import optimisation
 from . import hierarchical
-from . import evaluation
 from . import utils
+from . import datasets
 
 del os, sys, warnings, logging