+++ cpm/applications/__init__.py
@@ -1,2 +1,2 @@
-from . import DeltaNetwork
-
+from . import reinforcement_learning
+from . import signal_detection