+++ cpm/optimisation/free.py
@@ -2,6 +2,8 @@ from ..core.generators import generate_guesses
 from ..core.optimisers import objective, numerical_hessian, prepare_data
 from ..core.data import detailed_pandas_compiler, decompose
 from ..generators import Simulator, Wrapper
+from ..core.parallel import detect_cores, execute_parallel
+
 
 from scipy.optimize import minimize
 import numpy as np
@@ -31,6 +33,8 @@ class Minimize:
     cl : int
         The number of cores to use for parallel processing. Default is `None`. If `None`, the number of cores is set to 2.
         If `cl` is set to `None` and `parallel` is set to `True`, the number of cores is set to the number of cores available on the machine.
+    libraries : list, optional
+        The libraries to import for parallel processing for `ipyparallel` with the IPython kernel. Default is `["numpy", "pandas"]`
     ppt_identifier : str
         The key in the participant data dictionary that contains the participant identifier. Default is `None`. Returned in the optimization details.
     **kwargs : dict
@@ -53,6 +57,7 @@ class Minimize:
         method="Nelder-Mead",
         cl=None,
         parallel=False,
+        libraries=["numpy", "pandas"],
         prior=False,
         number_of_starts=1,
         ppt_identifier=None,
@@ -96,12 +101,13 @@ class Minimize:
         )
 
         self.__parallel__ = parallel
+        self.__libraries__ = libraries
         self.__current_guess__ = self.initial_guess[0]
 
         if cl is not None:
             self.cl = cl
         if cl is None and parallel:
-            self.cl = mp.cpu_count()
+            self.cl = detect_cores()
 
     def optimise(self):
         """
@@ -170,8 +176,14 @@ class Minimize:
             )
             self.__current_guess__ = self.initial_guess[i]
             if self.__parallel__:
-                with mp.Pool(self.cl) as pool:
-                    results = pool.map(__task, self.data)
+                results = execute_parallel(
+                    job=__task,
+                    data=self.data,
+                    method=None,
+                    cl=self.cl,
+                    pandas=self.__pandas__,
+                    libraries=self.__libraries__,
+                )
             else:
                 results = list(map(__task, self.data))